Refactoring looks good
