+++ js/global.js
@@ -1,0 +1,26 @@
+function currentTime() {
+	//returns current time in "[HH:MM:SS] " 24hr format (string)
+	var d = new Date();
+	return '[' + d.toTimeString().substr(0, 8) + '] ';
+}
+
+function randomNumBetween(min, max) {
+	//inclusive
+	return Math.floor(Math.random() * (max - min + 1)) + min;
+}
+
+function generateRoomCode() {
+	return (
+		'game' +
+		Math.random()
+			.toString(36)
+			.substr(2, 4)
+			.toUpperCase()
+	);
+}
+
+module.exports = {
+	currentTime,
+	randomNumBetween,
+	generateRoomCode,
+};
