+++ js/message.js
@@ -1,0 +1,15 @@
+const g = require('./global');
+
+class Message {
+	constructor(text, type = 'System', from = '', to = '') {
+		this.content =
+			type == 'Private'
+				? g.currentTime() + '(' + this.from + ' > ' + to + ') ' + text
+				: type == 'Public'
+				? g.currentTime() + from + ': ' + text
+				: g.currentTime() + text;
+		this.type = type;
+	}
+}
+
+module.exports = Message;
